# lib/api-zod/src/generated/types/createTestimonyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JCTM Digital Sanctuary API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTestimonyBody {
  name: string;
  /** @nullable */
  email?: string | null;
  content: string;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  category?: string | null;
}