# lib/api-zod/src/generated/types/donationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JCTM Digital Sanctuary API
 * OpenAPI spec version: 0.1.0
 */

export interface DonationResponse {
  reference: string;
  /** @nullable */
  authorizationUrl?: string | null;
  message: string;
}