# lib/api-zod/src/generated/types/givingLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JCTM Digital Sanctuary API
 * OpenAPI spec version: 0.1.0
 */

export interface GivingLog {
  id: number;
  /** @nullable */
  donorName?: string | null;
  /** @nullable */
  donorEmail?: string | null;
  amount: number;
  currency: string;
  /** @nullable */
  purpose?: string | null;
  reference: string;
  status: string;
  /** @nullable */
  paymentMethod?: string | null;
  createdAt: string;
}