# lib/api-zod/src/generated/types/livestreamStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JCTM Digital Sanctuary API
 * OpenAPI spec version: 0.1.0
 */

export interface LivestreamStatus {
  isLive: boolean;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  streamUrl?: string | null;
  /** @nullable */
  startedAt?: string | null;
}