# lib/api-zod/src/generated/types/sermonStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JCTM Digital Sanctuary API
 * OpenAPI spec version: 0.1.0
 */

export interface SermonStats {
  total: number;
  /** @nullable */
  latestDate?: string | null;
  /** @nullable */
  totalViews?: number | null;
}